requirements.txt
git+git://github.com/django/django.git#egg=Django
Markdown
psycopg2
dj-database-url
django-cbvpatterns
gunicorn
